PIQA 83 Removing circular dependencies (#103)
