+++ features/attribute/step_definitions/attribute-common-steps.ts
@@ -2,7 +2,7 @@ import assert from "assert";
 import { Given, Then } from "@cucumber/cucumber";
 import { z } from "zod";
 import { assertContextSchema, getToken } from "../../../utils/commons";
-import { TenantType } from "../../common-steps";
+import { TenantType } from "../../../utils/commons";
 import { AttributeKind } from "../../../api/models";
 import { dataPreparationService } from "../../../services/data-preparation.service";
