+++ features/catalog/step_definitions/e-service-risk-analysis-update.ts
@@ -7,7 +7,7 @@ import {
 } from "../../../utils/commons";
 import { dataPreparationService } from "../../../services/data-preparation.service";
 import { apiClient } from "../../../api";
-import { TenantType } from "../../common-steps";
+import { TenantType } from "../../../utils/commons";
 
 Given(
   "{string} ha già aggiunto un'analisi del rischio a quell'e-service",
