+++ features/purpose/step_definitions/purpose-creation-mode-deliver.ts
@@ -8,7 +8,7 @@ import {
   getToken,
 } from "../../../utils/commons";
 import { apiClient } from "../../../api";
-import { TenantType } from "../../common-steps";
+import { TenantType } from "../../../utils/commons";
 import {
   ESERVICE_DAILY_CALLS,
   dataPreparationService,
